// projects/oculr-ngx/src/lib/models/input-type.enum.ts
/*
 * @license
 * Copyright 2021 Progressive Casualty Insurance Company. All Rights Reserved.
 *
 * Use of this source code is governed by an MIT license that can be found in
 * the LICENSE file at https://github.com/Progressive/oculr-ngx/blob/main/LICENSE.md
 */

export enum InputType {
  checkbox = 'checkbox',
  date = 'date',
  email = 'email',
  number = 'number',
  password = 'password',
  radio = 'radio',
  search = 'search',
  select = 'select',
  tel = 'tel',
  text = 'text',
  textarea = 'textarea',
}