// projects/oculr-ngx/src/lib/models/interaction-detail.enum.ts
/*
 * @license
 * Copyright (c) 2025 Progressive Casualty Insurance Company. All Rights Reserved.
 *
 * Use of this source code is governed by an MIT license that can be found at
 * https://opensource.progressive.com/resources/license
*/

export enum InteractionDetail {
  mouse = 'mouse',
  keyboard = 'keyboard',
  touch = 'touch',
}